fix(tests): move mock to jest config
