+++ src/setupJest.tsx
@@ -631,3 +631,24 @@ jest.mock("@gorhom/bottom-sheet", () => ({
   __esModule: true,
   ...require("@gorhom/bottom-sheet/mock"),
 }))
+
+jest.mock("react-native-collapsible-tab-view", () => {
+  const React = require("react-native")
+  const MockedTabs = ({ name, children }: any) => (
+    <React.View>
+      <React.Text>{name}</React.Text>
+      {children}
+    </React.View>
+  )
+
+  return {
+    Tabs: {
+      Container: React.View,
+      Tab: MockedTabs,
+      FlatList: React.FlatList,
+      ScrollView: React.ScrollView,
+    },
+    useHeaderMeasurements: () => ({ height: { value: 0 } }),
+    useTabNameContext: () => "Tab",
+  }
+})
